Store UEID value in the instance field of created evidence, not in its own ECT
Fixes issue #6
To support evidence from multiple dice/spdm attesters, the `instance-id` field needs to be filled in with a unique identifier (within the namespace of the `class-id`.

For better document flow, I think it would be better to move the DiceUeid section before DiceTcbInfo. I did not make that move now for easier comparison of diffs.